+++ src/main/java/edu/metrostate/booknow/Models/User.java
@@ -2,17 +2,9 @@ package edu.metrostate.booknow.Models;
 
 public class User {
 
-    private final int userId;
-    private String usernName;
-    private String passWord;
+    private String username;
 
-    public User(int userId, String username, String password) {
-        this.userId = userId;
-        this.usernName = username;
-        this.passWord = password;
-    }
-
-    public String getUserName() {
-        return usernName;
+    public User(String username) {
+        this.username = username;
     }
 }